fixes part of the incremental tests
