+++ Gillian-C/examples/incremental/verification/test_add_proc/diff.txt.expected
@@ -1,8 +1,8 @@
 Changed procs:
-make_node
+listDispose
 New procs:
-find_min
+listCopy
 Deleted procs:
 <none>
 Transitive dependents:
-insert
+<none>
