--- Gillian-C/examples/incremental/verification/test_add_proc/src_after/bst.h
@@ -1,11 +1,0 @@
-#include <stdlib.h>
-#define FALSE 0
-#define TRUE 1
-
-typedef struct bstn {
-    int value;
-    struct bstn *left;
-    struct bstn *right;
-} BST;
-
-BST *make_node(int v);
