+++ Gillian-C/examples/incremental/verification/test_add_proc/src_after/sll_b.c
@@ -1,0 +1,30 @@
+#include "sll.h"
+
+/*@ spec listDispose(x) {
+  requires: list(#x, #alpha) * (x == #x)
+  ensures:  emp
+} */
+void listDispose(SLL *x) {
+    if (x == NULL) {
+        return;
+    } else {
+        listDispose(x->next);
+        free(x);
+        return;
+    };
+}
+
+/*@ spec listCopy(x) {
+  requires: list(#x, #alpha) * (x == #x)
+  ensures:  list(ret, #alpha) * list(#x, #alpha)
+} */
+SLL *listCopy(SLL *x) {
+    SLL *r;
+    if (x == NULL) {
+        r = NULL;
+    } else {
+        SLL *t = listCopy(x->next);
+        r = listPrependV(t, x->data);
+    };
+    return r;
+}
