+++ Gillian-C/examples/incremental/verification/test_add_proc/src_before/sll_b.c
@@ -1,0 +1,15 @@
+#include "sll.h"
+
+/*@ spec listDispose(x) {
+  requires: list(#x, #alpha) * (x == #x)
+  ensures:  emp
+} */
+void listDispose(SLL *x) {
+    if (x == NULL) {
+        return;
+    } else {
+        listDispose(x->next);
+        free(x);
+        return;
+    };
+}
