--- Gillian-C/examples/incremental/verification/test_change_header/src_after/bst.h
@@ -1,13 +1,0 @@
-#include <stdlib.h>
-#define FALSE 0
-#define TRUE 1
-
-typedef struct bstn {
-    int value;
-    struct bstn *left;
-    struct bstn *right;
-} BST;
-
-BST *make_node(int v);
-
-// Some change within the header file
