--- Gillian-C/examples/incremental/verification/test_change_header/src_before/bst_a.c
@@ -1,22 +1,0 @@
-#include "bst.h"
-
-/*@ spec insert(v, t) {
-  requires: (v == int(#vv)) * (t == #t) * BST(#t, #K)
-  ensures:  BST(ret, -u- (#K, -{ #vv }-))
-}*/
-BST *insert(int v, BST *t) {
-    BST *tmp;
-    if (t == NULL) {
-        return make_node(v);
-    }
-    if (v < t->value) {
-        tmp = insert(v, t->left);
-        __builtin_annot("unfold BST(tmp, -u- (-{ #vv }-, #KL))");
-        t->left = tmp;
-    } else if (v > t->value) {
-        tmp = insert(v, t->right);
-        __builtin_annot("unfold BST(tmp, -u- (-{ #vv }-, #KR))");
-        t->right = tmp;
-    }
-    return t;
-}
