+++ Gillian-C/examples/incremental/verification/test_change_header/src_before/sll.h
@@ -1,0 +1,10 @@
+#include <stdlib.h>
+#define FALSE 0
+#define TRUE 1
+
+typedef struct ln {
+    int data;
+    struct ln *next;
+} SLL;
+
+SLL* listPrependV(SLL *x, int v);
