+++ Gillian-C/examples/incremental/verification/test_change_header/src_before/sll_a.c
@@ -1,9 +1,4 @@
-#include <stdlib.h>
-
-typedef struct ln {
-    int data;
-    struct ln *next;
-} SLL;
+#include "sll.h"
 
 /*@ pred list(+p, alpha) {
   (p -m> struct ln { #head; #tail } * (alpha == #head::#beta)) *
@@ -101,52 +96,4 @@ int listLength(SLL *x) {
     } else {
         return 1 + listLength(x->next);
     };
-}
-
-/*@ spec listDispose(x) {
-  requires: list(#x, #alpha) * (x == #x)
-  ensures:  emp
-} */
-void listDispose(SLL *x) {
-    if (x == NULL) {
-        return;
-    } else {
-        listDispose(x->next);
-        free(x);
-        return;
-    };
-}
-
-/*@ spec listCopy(x) {
-  requires: list(#x, #alpha) * (x == #x)
-  ensures:  list(ret, #alpha) * list(#x, #alpha)
-} */
-SLL *listCopy(SLL *x) {
-    SLL *r;
-    if (x == NULL) {
-        r = NULL;
-    } else {
-        SLL *t = listCopy(x->next);
-        r = listPrependV(t, x->data);
-    };
-    return r;
-}
-
-/*@ spec listConcat(x, y) {
-  requires: list(#x, #alpha) * (x == #x) * list(#y, #beta) * (y == #y)
-  ensures:  list(ret, #alpha @ #beta)
-} */
-SLL *listConcat(SLL *x, SLL *y) {
-    SLL *r;
-    if (x == NULL) {
-        r = y;
-    } else {
-        SLL *c = listConcat(x->next, y);
-        __builtin_annot("assert [[bind #gamma]] list(c, #gamma)");
-        __builtin_annot("unfold list(c, #gamma)");
-        __builtin_annot("fold list(c, #gamma)");
-        x->next = c;
-        r = x;
-    };
-    return r;
 }
